+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## Cockpit File Sharing 2.0.2-1
+## Cockpit File Sharing 2.1.0-1
 
-* Fixed small bugs
+* Added more error catching and better descriptions